+++ ChangeLog.md
@@ -1,6 +1,10 @@
 # WALib - common library Copyright (c) 2020-2021
-Author: Walery Wysotsky <nkrzi@wysotsky.info>
+Author: Walery Wysotsky <dev@wysotsky.info>
+
+## [0.4.0] - 2021-03-08
+### Added
+- ResourceUtils class - various resource manipulation
 
 ## [0.3.0] - 2021-01-30
 ### Added
-- ClasspathResourceUtil class
+- ResourceLoader class